pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Task Manager application"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "django>=5.0,<6.0",
    "python-dotenv>=1.0",
    "dj-database-url>=2.1",
    "gunicorn>=21.2",
    "whitenoise>=6.6",
    "django-bootstrap5>=24.3",
    "django-filter>=24.3",
    "rollbar>=0.16.3",
    "psycopg[binary]>=3.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=5.0",
    "ruff>=0.6.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "task_manager.settings"
python_files = ["tests.py", "test_*.py", "*_test.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_manager", "task_manager_app"]